refactor: extract delete, extend, download workflows and published-ad fetch into focused modules
- Create published_ads.py with fetch_published_ads() and PublishedAdsFetchIncompleteError
- Create delete_flow.py with delete_ads() and delete_ad()
- Create extend_flow.py with extend_ads() and _extend_ad()
- Create download_flow.py with download_ads(), resolve_download_dir(), and _download_ad_with_resolved_state()
- Promote _timeout and _navigate_paginated_ad_overview to public in WebScrapingMixin
- Shrink _init_.py from 2865 to 2391 lines (-16.5%)
- Create focused test files: test_published_ads.py, test_delete_flow.py, test_extend_flow.py, test_download_flow.py (79 tests)
- Update translations and fix all patch targets
