+++ tests/unit/test_init.py
@@ -28,7 +28,7 @@ from kleinanzeigen_bot.model.config_model import (
     PublishingConfig,
 )
 from kleinanzeigen_bot.utils import xdg_paths
-from kleinanzeigen_bot.utils.exceptions import CategoryResolutionError, PublishedAdsFetchIncompleteError, PublishSubmissionUncertainError
+from kleinanzeigen_bot.utils.exceptions import CategoryResolutionError, PublishSubmissionUncertainError
 from kleinanzeigen_bot.utils.web_scraping_mixin import By, Element
 
 
@@ -159,7 +159,9 @@ class TestKleinanzeigenBotInitialization:
             patch.object(test_bot, "load_ads", return_value = []),
             patch.object(test_bot, "create_browser_session", new_callable = AsyncMock),
             patch.object(test_bot, "login", new_callable = AsyncMock),
-            patch.object(test_bot, "download_ads", new_callable = AsyncMock),
+            patch("kleinanzeigen_bot.download_flow.download_ads", new_callable = AsyncMock),
+
+
             patch.object(test_bot, "close_browser_session"),
             patch("kleinanzeigen_bot.UpdateChecker", DummyUpdateChecker),
         ):
@@ -167,472 +169,6 @@ class TestKleinanzeigenBotInitialization:
 
         expected_state_path = (tmp_path / "config.yaml").resolve().parent / ".temp" / "update_check_state.json"
         assert update_checker_calls == [(test_bot.config, expected_state_path)]
-
-    @pytest.mark.asyncio
-    async def test_download_ads_passes_download_dir_and_published_ads(self, test_bot:KleinanzeigenBot, tmp_path:Path) -> None:
-        """Ensure download_ads wires resolved download_dir and published_ads_by_id into AdExtractor."""
-        config_file = tmp_path / "config.yaml"
-        test_bot.workspace = xdg_paths.Workspace.for_config(config_file, "kleinanzeigen-bot")
-        test_bot.config_file_path = str(config_file)
-        test_bot.ads_selector = "all"
-        test_bot.browser = MagicMock()
-
-        extractor_mock = MagicMock()
-        extractor_mock.extract_own_ads_urls = AsyncMock(return_value = [])
-
-        mock_published_ads = [{"id": 123, "buyNowEligible": True}, {"id": 456, "buyNowEligible": False}]
-
-        with (
-            patch.object(test_bot, "_fetch_published_ads", new_callable = AsyncMock, return_value = mock_published_ads),
-            patch("kleinanzeigen_bot.extract.AdExtractor", return_value = extractor_mock) as mock_extractor,
-        ):
-            await test_bot.download_ads()
-
-        # Verify published_ads_by_id is built correctly and passed to extractor
-        mock_extractor.assert_called_once_with(
-            test_bot.browser,
-            test_bot.config,
-            test_bot.workspace.download_dir,
-            published_ads_by_id = {123: mock_published_ads[0], 456: mock_published_ads[1]},
-        )
-
-    def test_resolve_download_dir_uses_workspace_default_for_literal_default(self, test_bot:KleinanzeigenBot, tmp_path:Path) -> None:
-        config_file = tmp_path / "config.yaml"
-        test_bot.workspace = xdg_paths.Workspace.for_config(config_file, "kleinanzeigen-bot")
-        test_bot.config_file_path = str(config_file)
-        test_bot.browser = cast(Any, None)
-
-        assert test_bot._resolve_download_dir() == test_bot.workspace.download_dir
-
-    def test_resolve_download_dir_uses_config_relative_path(self, test_bot:KleinanzeigenBot, tmp_path:Path) -> None:
-        config_file = tmp_path / "config.yaml"
-        test_bot.workspace = xdg_paths.Workspace.for_config(config_file, "kleinanzeigen-bot")
-        test_bot.config_file_path = str(config_file)
-        test_bot.browser = cast(Any, None)
-        test_bot.config.download.dir = "./my-ads"
-
-        assert test_bot._resolve_download_dir() == (tmp_path / "my-ads").resolve()
-
-    def test_resolve_download_dir_uses_absolute_path(self, test_bot:KleinanzeigenBot, tmp_path:Path) -> None:
-        config_file = tmp_path / "config.yaml"
-        test_bot.workspace = xdg_paths.Workspace.for_config(config_file, "kleinanzeigen-bot")
-        test_bot.config_file_path = str(config_file)
-        test_bot.browser = cast(Any, None)
-        test_bot.config.download.dir = str((tmp_path / "absolute-target").resolve())
-
-        assert test_bot._resolve_download_dir() == (tmp_path / "absolute-target").resolve()
-
-    @pytest.mark.asyncio
-    async def test_download_ads_uses_configured_relative_download_dir(self, test_bot:KleinanzeigenBot, tmp_path:Path) -> None:
-        config_file = tmp_path / "config.yaml"
-        test_bot.workspace = xdg_paths.Workspace.for_config(config_file, "kleinanzeigen-bot")
-        test_bot.config_file_path = str(config_file)
-        test_bot.config.download.dir = "./ads"
-        test_bot.ads_selector = "all"
-        test_bot.browser = MagicMock()
-
-        extractor_mock = MagicMock()
-        extractor_mock.extract_own_ads_urls = AsyncMock(return_value = [])
-
-        with (
-            patch.object(test_bot, "_fetch_published_ads", new_callable = AsyncMock, return_value = []),
-            patch("kleinanzeigen_bot.extract.AdExtractor", return_value = extractor_mock) as mock_extractor,
-        ):
-            await test_bot.download_ads()
-
-        mock_extractor.assert_called_once()
-        assert mock_extractor.call_args.args[2] == (tmp_path / "ads").resolve()
-
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize(
-        "scenario",
-        [
-            {
-                "published_ads": [{"id": 123, "state": "active"}],
-                "expected_active": True,
-            },
-            {
-                "published_ads": [{"id": 999, "state": "active"}],
-                "expected_active": False,
-            },
-            {
-                "published_ads": [{"id": 123, "state": "inactive"}],
-                "expected_active": False,
-            },
-            {
-                "published_ads": [{"id": 123, "state": "paused"}],
-                "expected_active": False,
-            },
-        ],
-    )
-    async def test_download_ads_numeric_selector_resolves_and_passes_active_state(
-        self,
-        test_bot:KleinanzeigenBot,
-        tmp_path:Path,
-        scenario:dict[str, Any],
-    ) -> None:
-        published_ads = scenario["published_ads"]
-        expected_active = scenario["expected_active"]
-
-        test_bot.workspace = xdg_paths.Workspace.for_config(tmp_path / "config.yaml", "kleinanzeigen-bot")
-        test_bot.ads_selector = "123"
-        test_bot.browser = MagicMock()
-
-        extractor_mock = MagicMock()
-        extractor_mock.navigate_to_ad_page = AsyncMock(return_value = True)
-        extractor_mock.download_ad = AsyncMock()
-
-        with (
-            patch.object(test_bot, "_fetch_published_ads", new_callable = AsyncMock, return_value = published_ads) as mock_fetch_published_ads,
-            patch("kleinanzeigen_bot.extract.AdExtractor", return_value = extractor_mock),
-        ):
-            await test_bot.download_ads()
-
-        mock_fetch_published_ads.assert_awaited_once_with(strict = True)
-
-        extractor_mock.download_ad.assert_awaited_once_with(123, active = expected_active)
-
-    @pytest.mark.asyncio
-    async def test_download_ads_numeric_selector_fails_when_published_ads_fetch_incomplete(self, test_bot:KleinanzeigenBot, tmp_path:Path) -> None:
-        test_bot.workspace = xdg_paths.Workspace.for_config(tmp_path / "config.yaml", "kleinanzeigen-bot")
-        test_bot.ads_selector = "123"
-        test_bot.browser = MagicMock()
-
-        with (
-            patch.object(
-                test_bot,
-                "_fetch_published_ads",
-                new_callable = AsyncMock,
-                side_effect = PublishedAdsFetchIncompleteError("incomplete fetch"),
-            ),
-            patch("kleinanzeigen_bot.extract.AdExtractor") as mock_extractor,
-            pytest.raises(PublishedAdsFetchIncompleteError, match = "incomplete fetch"),
-        ):
-            await test_bot.download_ads()
-
-        mock_extractor.assert_not_called()
-
-    @pytest.mark.asyncio
-    async def test_download_ads_all_selector_uses_tolerant_published_ads_fetch(self, test_bot:KleinanzeigenBot, tmp_path:Path) -> None:
-        test_bot.workspace = xdg_paths.Workspace.for_config(tmp_path / "config.yaml", "kleinanzeigen-bot")
-        test_bot.ads_selector = "all"
-        test_bot.browser = MagicMock()
-
-        extractor_mock = MagicMock()
-        extractor_mock.extract_own_ads_urls = AsyncMock(return_value = [])
-
-        with (
-            patch.object(test_bot, "_fetch_published_ads", new_callable = AsyncMock, return_value = []) as mock_fetch_published_ads,
-            patch("kleinanzeigen_bot.extract.AdExtractor", return_value = extractor_mock),
-        ):
-            await test_bot.download_ads()
-
-        mock_fetch_published_ads.assert_awaited_once_with(strict = False)
-
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize(
-        "scenario",
-        [
-            {
-                "name": "active ad",
-                "published_ads": [{"id": 123, "state": "active"}],
-                "expected_active": True,
-                "expect_ownership_warning": False,
-            },
-            {
-                "name": "inactive ad",
-                "published_ads": [{"id": 123, "state": "inactive"}],
-                "expected_active": False,
-                "expect_ownership_warning": False,
-            },
-            {
-                "name": "paused ad",
-                "published_ads": [{"id": 123, "state": "paused"}],
-                "expected_active": False,
-                "expect_ownership_warning": False,
-            },
-            {
-                "name": "ad not in published profile",
-                "published_ads": [{"id": 999, "state": "active"}],  # Different ID
-                "expected_active": False,
-                "expect_ownership_warning": True,
-            },
-        ],
-        ids = lambda s: s["name"],
-    )
-    async def test_download_ads_all_selector_resolves_and_passes_active_state(
-        self,
-        test_bot:KleinanzeigenBot,
-        tmp_path:Path,
-        caplog:pytest.LogCaptureFixture,
-        scenario:dict[str, Any],
-    ) -> None:
-        """Test that --ads=all resolves and passes correct active state to download_ad."""
-        # Setup
-        test_bot.workspace = xdg_paths.Workspace.for_config(tmp_path / "config.yaml", "kleinanzeigen-bot")
-        test_bot.ads_selector = "all"
-        test_bot.browser = MagicMock()
-
-        extractor_mock = MagicMock()
-        extractor_mock.extract_own_ads_urls = AsyncMock(return_value = ["https://www.kleinanzeigen.de/s-anzeige/test-ad/123-234-5678"])
-        extractor_mock.extract_ad_id_from_ad_url = MagicMock(return_value = 123)
-        extractor_mock.navigate_to_ad_page = AsyncMock(return_value = True)
-        extractor_mock.download_ad = AsyncMock()
-
-        caplog.set_level(logging.WARNING)
-
-        with (
-            patch.object(test_bot, "_fetch_published_ads", new_callable = AsyncMock, return_value = scenario["published_ads"]) as mock_fetch_published_ads,
-            patch("kleinanzeigen_bot.extract.AdExtractor", return_value = extractor_mock),
-        ):
-            await test_bot.download_ads()
-
-        # Verify published ads fetched with strict=False (tolerant mode for "all")
-        mock_fetch_published_ads.assert_awaited_once_with(strict = False)
-
-        # Verify download_ad called with correct active parameter
-        extractor_mock.download_ad.assert_awaited_once_with(123, active = scenario["expected_active"])
-
-        # Verify ownership warning only when expected
-        ownership_warnings = [msg for msg in caplog.messages if "found in overview but not in published profile" in msg]
-        if scenario["expect_ownership_warning"]:
-            assert len(ownership_warnings) == 1
-        else:
-            assert len(ownership_warnings) == 0
-
-    @pytest.mark.asyncio
-    async def test_download_ads_all_selector_skips_invalid_ad_id(
-        self,
-        test_bot:KleinanzeigenBot,
-        tmp_path:Path,
-    ) -> None:
-        """Test that --ads=all skips ads with invalid URL parsing (ad_id=-1)."""
-        test_bot.workspace = xdg_paths.Workspace.for_config(tmp_path / "config.yaml", "kleinanzeigen-bot")
-        test_bot.ads_selector = "all"
-        test_bot.browser = MagicMock()
-
-        extractor_mock = MagicMock()
-        extractor_mock.extract_own_ads_urls = AsyncMock(return_value = ["https://www.kleinanzeigen.de/s-anzeige/test/invalid-url"])
-        extractor_mock.extract_ad_id_from_ad_url = MagicMock(return_value = -1)  # URL parsing failed
-        extractor_mock.navigate_to_ad_page = AsyncMock(return_value = True)
-        extractor_mock.download_ad = AsyncMock()
-
-        with (
-            patch.object(test_bot, "_fetch_published_ads", new_callable = AsyncMock, return_value = []),
-            patch("kleinanzeigen_bot.extract.AdExtractor", return_value = extractor_mock),
-        ):
-            await test_bot.download_ads()
-
-        # Verify download_ad was NOT called for invalid ad_id
-        extractor_mock.download_ad.assert_not_called()
-
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize(
-        "scenario",
-        [
-            {
-                "name": "new active ad",
-                "published_ads": [{"id": 999, "state": "active"}],
-                "saved_ad_ids": [123, 456],  # 999 is not saved, so it's "new"
-                "expected_active": True,
-            },
-            {
-                "name": "new inactive ad",
-                "published_ads": [{"id": 999, "state": "inactive"}],
-                "saved_ad_ids": [123, 456],  # 999 is not saved, so it's "new"
-                "expected_active": False,
-            },
-            {
-                "name": "new paused ad",
-                "published_ads": [{"id": 999, "state": "paused"}],
-                "saved_ad_ids": [123, 456],  # 999 is not saved, so it's "new"
-                "expected_active": False,
-            },
-        ],
-        ids = lambda s: s["name"],
-    )
-    async def test_download_ads_new_selector_resolves_and_passes_active_state(
-        self,
-        test_bot:KleinanzeigenBot,
-        tmp_path:Path,
-        scenario:dict[str, Any],
-    ) -> None:
-        """Test that --ads=new resolves and passes correct active state to download_ad."""
-        test_bot.workspace = xdg_paths.Workspace.for_config(tmp_path / "config.yaml", "kleinanzeigen-bot")
-        test_bot.ads_selector = "new"
-        test_bot.browser = MagicMock()
-
-        extractor_mock = MagicMock()
-        extractor_mock.extract_own_ads_urls = AsyncMock(return_value = ["https://www.kleinanzeigen.de/s-anzeige/test-ad/999-234-5678"])
-        extractor_mock.extract_ad_id_from_ad_url = MagicMock(return_value = 999)
-        extractor_mock.navigate_to_ad_page = AsyncMock(return_value = True)
-        extractor_mock.download_ad = AsyncMock()
-
-        # Mock load_ads to return the saved_ad_ids
-        saved_ads:list[tuple[str, MagicMock, dict[str, Any]]] = [
-            (
-                f"ad_{ad_id}.yaml",
-                MagicMock(spec = Ad, id = ad_id),
-                {},
-            )
-            for ad_id in scenario["saved_ad_ids"]
-        ]
-
-        with (
-            patch.object(test_bot, "_fetch_published_ads", new_callable = AsyncMock, return_value = scenario["published_ads"]) as mock_fetch_published_ads,
-            patch("kleinanzeigen_bot.extract.AdExtractor", return_value = extractor_mock),
-            patch.object(test_bot, "load_ads", return_value = saved_ads),
-        ):
-            await test_bot.download_ads()
-
-        # Verify published ads fetched with strict=False (tolerant mode for "new")
-        mock_fetch_published_ads.assert_awaited_once_with(strict = False)
-
-        # Verify download_ad called with correct active parameter
-        extractor_mock.download_ad.assert_awaited_once_with(999, active = scenario["expected_active"])
-
-    @pytest.mark.asyncio
-    async def test_download_ads_new_selector_skips_already_saved(
-        self,
-        test_bot:KleinanzeigenBot,
-        tmp_path:Path,
-    ) -> None:
-        """Test that --ads=new skips already-saved ads (existing behavior unchanged)."""
-        test_bot.workspace = xdg_paths.Workspace.for_config(tmp_path / "config.yaml", "kleinanzeigen-bot")
-        test_bot.ads_selector = "new"
-        test_bot.browser = MagicMock()
-
-        extractor_mock = MagicMock()
-        extractor_mock.extract_own_ads_urls = AsyncMock(return_value = ["https://www.kleinanzeigen.de/s-anzeige/test-ad/123-234-5678"])
-        extractor_mock.extract_ad_id_from_ad_url = MagicMock(return_value = 123)
-        extractor_mock.navigate_to_ad_page = AsyncMock(return_value = True)
-        extractor_mock.download_ad = AsyncMock()
-
-        # Mock load_ads to return ad 123 as already saved
-        saved_ads:list[tuple[str, MagicMock, dict[str, Any]]] = [("ad_123.yaml", MagicMock(spec = Ad, id = 123), {})]
-
-        with (
-            patch.object(test_bot, "_fetch_published_ads", new_callable = AsyncMock, return_value = []),
-            patch("kleinanzeigen_bot.extract.AdExtractor", return_value = extractor_mock),
-            patch.object(test_bot, "load_ads", return_value = saved_ads),
-        ):
-            await test_bot.download_ads()
-
-        # Verify download_ad was NOT called for already-saved ad
-        extractor_mock.download_ad.assert_not_called()
-
-    @pytest.mark.asyncio
-    async def test_download_ads_new_selector_skips_invalid_ad_id(
-        self,
-        test_bot:KleinanzeigenBot,
-        tmp_path:Path,
-    ) -> None:
-        """Test that --ads=new skips ads with invalid URL parsing (ad_id=-1)."""
-        test_bot.workspace = xdg_paths.Workspace.for_config(tmp_path / "config.yaml", "kleinanzeigen-bot")
-        test_bot.ads_selector = "new"
-        test_bot.browser = MagicMock()
-
-        extractor_mock = MagicMock()
-        extractor_mock.extract_own_ads_urls = AsyncMock(return_value = ["https://www.kleinanzeigen.de/s-anzeige/test/invalid-url"])
-        extractor_mock.extract_ad_id_from_ad_url = MagicMock(return_value = -1)  # URL parsing failed
-        extractor_mock.navigate_to_ad_page = AsyncMock(return_value = True)
-        extractor_mock.download_ad = AsyncMock()
-
-        with (
-            patch.object(test_bot, "_fetch_published_ads", new_callable = AsyncMock, return_value = []),
-            patch("kleinanzeigen_bot.extract.AdExtractor", return_value = extractor_mock),
-            patch.object(test_bot, "load_ads", return_value = []),
-        ):
-            await test_bot.download_ads()
-
-        # Verify download_ad was NOT called for invalid ad_id
-        extractor_mock.download_ad.assert_not_called()
-
-    @pytest.mark.asyncio
-    async def test_download_ads_new_selector_passes_inactive_for_ad_not_in_published_profile(
-        self,
-        test_bot:KleinanzeigenBot,
-        tmp_path:Path,
-    ) -> None:
-        """Test that --ads=new passes active=False when ad is not in published profile."""
-        test_bot.workspace = xdg_paths.Workspace.for_config(tmp_path / "config.yaml", "kleinanzeigen-bot")
-        test_bot.ads_selector = "new"
-        test_bot.browser = MagicMock()
-
-        extractor_mock = MagicMock()
-        extractor_mock.extract_own_ads_urls = AsyncMock(return_value = ["https://www.kleinanzeigen.de/s-anzeige/test-ad/999-234-5678"])
-        extractor_mock.extract_ad_id_from_ad_url = MagicMock(return_value = 999)
-        extractor_mock.navigate_to_ad_page = AsyncMock(return_value = True)
-        extractor_mock.download_ad = AsyncMock()
-
-        # Mock load_ads to return different saved ads (999 is new but not in published profile)
-        saved_ads:list[tuple[str, MagicMock, dict[str, Any]]] = [("ad_123.yaml", MagicMock(spec = Ad, id = 123), {})]
-
-        with (
-            patch.object(test_bot, "_fetch_published_ads", new_callable = AsyncMock, return_value = []),
-            patch("kleinanzeigen_bot.extract.AdExtractor", return_value = extractor_mock),
-            patch.object(test_bot, "load_ads", return_value = saved_ads),
-        ):
-            await test_bot.download_ads()
-
-        # Verify download_ad was called with active=False (not in profile)
-        extractor_mock.download_ad.assert_awaited_once_with(999, active = False)
-
-    @pytest.mark.asyncio
-    async def test_download_ads_all_selector_skips_when_navigation_fails(
-        self,
-        test_bot:KleinanzeigenBot,
-        tmp_path:Path,
-    ) -> None:
-        """Test that --ads=all skips download when navigate_to_ad_page returns False."""
-        test_bot.workspace = xdg_paths.Workspace.for_config(tmp_path / "config.yaml", "kleinanzeigen-bot")
-        test_bot.ads_selector = "all"
-        test_bot.browser = MagicMock()
-
-        extractor_mock = MagicMock()
-        extractor_mock.extract_own_ads_urls = AsyncMock(return_value = ["https://www.kleinanzeigen.de/s-anzeige/test/123"])
-        extractor_mock.extract_ad_id_from_ad_url = MagicMock(return_value = 123)
-        extractor_mock.navigate_to_ad_page = AsyncMock(return_value = False)  # Navigation fails
-        extractor_mock.download_ad = AsyncMock()
-
-        with (
-            patch.object(test_bot, "_fetch_published_ads", AsyncMock(return_value = [{"id": 123, "state": "active"}])),
-            patch("kleinanzeigen_bot.extract.AdExtractor", return_value = extractor_mock),
-        ):
-            await test_bot.download_ads()
-
-        # Verify download_ad was NOT called when navigation fails
-        extractor_mock.download_ad.assert_not_called()
-
-    @pytest.mark.asyncio
-    @pytest.mark.parametrize("state_value", ["", "deleted", "expired", "pending", "draft", None])
-    async def test_download_ads_all_selector_treats_unexpected_states_as_inactive(
-        self,
-        test_bot:KleinanzeigenBot,
-        tmp_path:Path,
-        state_value:str | None,
-    ) -> None:
-        """Test that unexpected state values are treated as inactive."""
-        test_bot.workspace = xdg_paths.Workspace.for_config(tmp_path / "config.yaml", "kleinanzeigen-bot")
-        test_bot.ads_selector = "all"
-        test_bot.browser = MagicMock()
-
-        published_ads = [{"id": 123, "state": state_value}] if state_value is not None else [{"id": 123}]
-
-        extractor_mock = MagicMock()
-        extractor_mock.extract_own_ads_urls = AsyncMock(return_value = ["https://www.kleinanzeigen.de/s-anzeige/test/123"])
-        extractor_mock.extract_ad_id_from_ad_url = MagicMock(return_value = 123)
-        extractor_mock.navigate_to_ad_page = AsyncMock(return_value = True)
-        extractor_mock.download_ad = AsyncMock()
-
-        with (
-            patch.object(test_bot, "_fetch_published_ads", AsyncMock(return_value = published_ads)),
-            patch("kleinanzeigen_bot.extract.AdExtractor", return_value = extractor_mock),
-        ):
-            await test_bot.download_ads()
-
-        # All non-"active" states should result in active=False
-        extractor_mock.download_ad.assert_awaited_once_with(123, active = False)
 
 
 class TestKleinanzeigenBotAuthentication:
@@ -693,7 +229,7 @@ class TestKleinanzeigenBotAuthentication:
             assert await test_bot.is_logged_in() is True
 
         group_text.assert_awaited()
-        assert any(call.kwargs.get("timeout") == test_bot._timeout("login_detection") for call in group_text.await_args_list)
+        assert any(call.kwargs.get("timeout") == test_bot.timeout("login_detection") for call in group_text.await_args_list)
 
     @pytest.mark.asyncio
     async def test_is_logged_in_runs_full_selector_group_before_cta_precheck(self, test_bot:KleinanzeigenBot) -> None:
@@ -1237,7 +773,7 @@ class TestKleinanzeigenBotAuthentication:
     async def test_click_gdpr_banner_uses_quick_dom_timeout_and_clicks_found_element(self, test_bot:KleinanzeigenBot) -> None:
         mock_element = AsyncMock()
         with (
-            patch.object(test_bot, "_timeout", return_value = 1.25) as mock_timeout,
+            patch.object(test_bot, "timeout", return_value = 1.25) as mock_timeout,
             patch.object(test_bot, "web_probe", new_callable = AsyncMock, return_value = mock_element) as mock_probe,
             patch.object(test_bot, "web_sleep", new_callable = AsyncMock) as mock_sleep,
         ):
@@ -1253,7 +789,7 @@ class TestKleinanzeigenBotAuthentication:
     @pytest.mark.asyncio
     async def test_click_gdpr_banner_does_nothing_when_banner_absent(self, test_bot:KleinanzeigenBot) -> None:
         with (
-            patch.object(test_bot, "_timeout", return_value = 1.25),
+            patch.object(test_bot, "timeout", return_value = 1.25),
             patch.object(test_bot, "web_probe", new_callable = AsyncMock, return_value = None) as mock_probe,
             patch.object(test_bot, "web_sleep", new_callable = AsyncMock) as mock_sleep,
         ):
@@ -1266,7 +802,7 @@ class TestKleinanzeigenBotAuthentication:
     async def test_dismiss_consent_banner_clicks_when_present(self, test_bot:KleinanzeigenBot) -> None:
         mock_element = AsyncMock()
         with (
-            patch.object(test_bot, "_timeout", return_value = 2.0) as mock_timeout,
+            patch.object(test_bot, "timeout", return_value = 2.0) as mock_timeout,
             patch.object(test_bot, "web_probe", new_callable = AsyncMock, return_value = mock_element) as mock_probe,
             patch.object(test_bot, "web_sleep", new_callable = AsyncMock) as mock_sleep,
         ):
@@ -1282,7 +818,7 @@ class TestKleinanzeigenBotAuthentication:
     @pytest.mark.asyncio
     async def test_dismiss_consent_banner_does_nothing_when_absent(self, test_bot:KleinanzeigenBot) -> None:
         with (
-            patch.object(test_bot, "_timeout", return_value = 2.0),
+            patch.object(test_bot, "timeout", return_value = 2.0),
             patch.object(test_bot, "web_probe", new_callable = AsyncMock, return_value = None) as mock_probe,
             patch.object(test_bot, "web_sleep", new_callable = AsyncMock) as mock_sleep,
         ):
@@ -1295,7 +831,7 @@ class TestKleinanzeigenBotAuthentication:
     async def test_check_sms_verification_prompts_user_when_detected(self, test_bot:KleinanzeigenBot) -> None:
         mock_element = MagicMock()
         with (
-            patch.object(test_bot, "_timeout", return_value = 3.0) as mock_timeout,
+            patch.object(test_bot, "timeout", return_value = 3.0) as mock_timeout,
             patch.object(test_bot, "web_probe", new_callable = AsyncMock, return_value = mock_element) as mock_probe,
             patch("kleinanzeigen_bot.ainput", new_callable = AsyncMock) as mock_ainput,
         ):
@@ -1310,7 +846,7 @@ class TestKleinanzeigenBotAuthentication:
     @pytest.mark.asyncio
     async def test_check_sms_verification_returns_silently_when_absent(self, test_bot:KleinanzeigenBot) -> None:
         with (
-            patch.object(test_bot, "_timeout", return_value = 3.0),
+            patch.object(test_bot, "timeout", return_value = 3.0),
             patch.object(test_bot, "web_probe", new_callable = AsyncMock, return_value = None) as mock_probe,
             patch("kleinanzeigen_bot.ainput", new_callable = AsyncMock) as mock_ainput,
         ):
@@ -1323,7 +859,7 @@ class TestKleinanzeigenBotAuthentication:
     async def test_check_email_verification_prompts_user_when_detected(self, test_bot:KleinanzeigenBot) -> None:
         mock_element = MagicMock()
         with (
-            patch.object(test_bot, "_timeout", return_value = 4.0) as mock_timeout,
+            patch.object(test_bot, "timeout", return_value = 4.0) as mock_timeout,
             patch.object(test_bot, "web_probe", new_callable = AsyncMock, return_value = mock_element) as mock_probe,
             patch("kleinanzeigen_bot.ainput", new_callable = AsyncMock) as mock_ainput,
         ):
@@ -1338,7 +874,7 @@ class TestKleinanzeigenBotAuthentication:
     @pytest.mark.asyncio
     async def test_check_email_verification_returns_silently_when_absent(self, test_bot:KleinanzeigenBot) -> None:
         with (
-            patch.object(test_bot, "_timeout", return_value = 4.0),
+            patch.object(test_bot, "timeout", return_value = 4.0),
             patch.object(test_bot, "web_probe", new_callable = AsyncMock, return_value = None) as mock_probe,
             patch("kleinanzeigen_bot.ainput", new_callable = AsyncMock) as mock_ainput,
         ):
@@ -1509,7 +1045,7 @@ class TestKleinanzeigenBotBasics:
             patch.object(test_bot, "web_request", new_callable = AsyncMock, return_value = {"content": json.dumps(payload)}) as web_request_mock,
             patch.object(test_bot, "publish_ad", new_callable = AsyncMock) as publish_ad_mock,
             patch.object(test_bot, "web_await", new_callable = AsyncMock, return_value = True) as web_await_mock,
-            patch.object(test_bot, "delete_ad", new_callable = AsyncMock) as delete_ad_mock,
+            patch("kleinanzeigen_bot.delete_flow.delete_ad", new_callable = AsyncMock) as delete_ad_mock,
         ):
             await test_bot.publish_ads(ad_cfgs)
 
@@ -1518,7 +1054,12 @@ class TestKleinanzeigenBotBasics:
             web_request_mock.assert_awaited_once_with(expected_url)
             publish_ad_mock.assert_awaited_once_with("ad.yaml", ad_cfgs[0][1], {}, [], AdUpdateStrategy.REPLACE)
             web_await_mock.assert_awaited_once()
-            delete_ad_mock.assert_awaited_once_with(ad_cfgs[0][1], [], delete_old_ads_by_title = False)
+            delete_ad_mock.assert_awaited_once_with(
+                web = test_bot, root_url = test_bot.root_url,
+                ad_cfg = ad_cfgs[0][1],
+                published_ads_list = [],
+                delete_old_ads_by_title = False,
+            )
 
     @pytest.mark.asyncio
     async def test_publish_ads_uses_millisecond_retry_delay_on_retryable_failure(
@@ -2309,7 +1850,7 @@ class TestKleinanzeigenBotAdOperations:
     @pytest.mark.asyncio
     async def test_run_download_command_default_selector(self, test_bot:KleinanzeigenBot, mock_config_setup:None) -> None:  # pylint: disable=unused-argument
         """Test running download command with default selector."""
-        with patch.object(test_bot, "download_ads", new_callable = AsyncMock):
+        with patch("kleinanzeigen_bot.download_flow.download_ads", new_callable = AsyncMock):
             await test_bot.run(["script.py", "download"])
             assert test_bot.ads_selector == "new"
 
@@ -2335,7 +1876,7 @@ class TestKleinanzeigenBotAdManagement:
     async def test_download_ads_with_specific_ids(self, test_bot:KleinanzeigenBot, mock_config_setup:None) -> None:  # pylint: disable=unused-argument
         """Test downloading ads with specific IDs."""
         test_bot.ads_selector = "123,456"
-        with patch.object(test_bot, "download_ads", new_callable = AsyncMock):
+        with patch("kleinanzeigen_bot.download_flow.download_ads", new_callable = AsyncMock):
             await test_bot.run(["script.py", "download", "--ads=123,456"])
             assert test_bot.ads_selector == "123,456"
 
@@ -2366,209 +1907,6 @@ class TestKleinanzeigenBotAdManagement:
         with pytest.raises(SystemExit) as exc_info:
             await test_bot.run(["script.py", "extend", "--ads=invalid"])
         assert exc_info.value.code == 2
-
-
-class TestKleinanzeigenBotAdDeletion:
-    """Tests for ad deletion functionality."""
-
-    @pytest.mark.asyncio
-    async def test_delete_ad_by_title_match_succeeds(
-        self,
-        test_bot:KleinanzeigenBot,
-        minimal_ad_config:dict[str, Any],
-    ) -> None:
-        """When title matches a published ad and server returns 200, should return True and clear id."""
-        ad_cfg = Ad.model_validate(minimal_ad_config | {"title": "Test Title", "id": None})
-        published_ads = [{"title": "Test Title", "id": 67890}, {"title": "Other Title", "id": 11111}]
-
-        with (
-            patch.object(test_bot, "web_open", new_callable = AsyncMock),
-            patch.object(test_bot, "web_find", new_callable = AsyncMock) as mock_find,
-            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
-            patch.object(test_bot, "web_request", new_callable = AsyncMock,
-                         return_value = {"statusCode": 200, "statusMessage": "OK", "content": "{}"}),
-        ):
-            mock_find.return_value.attrs = {"content": "some-token"}
-            result = await test_bot.delete_ad(ad_cfg, published_ads, delete_old_ads_by_title = True)
-
-        assert result is True
-        assert ad_cfg.id is None
-
-    @pytest.mark.asyncio
-    async def test_delete_ad_by_id_succeeds(self, test_bot:KleinanzeigenBot, minimal_ad_config:dict[str, Any]) -> None:
-        """When ad has an ID and server returns 200, should return True and clear id."""
-        ad_cfg = Ad.model_validate(minimal_ad_config | {"id": 12345})
-        published_ads = [{"title": "Different Title", "id": 12345}, {"title": "Other Title", "id": 11111}]
-
-        with (
-            patch.object(test_bot, "web_open", new_callable = AsyncMock),
-            patch.object(test_bot, "web_find", new_callable = AsyncMock) as mock_find,
-            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
-            patch.object(test_bot, "web_request", new_callable = AsyncMock,
-                         return_value = {"statusCode": 200, "statusMessage": "OK", "content": "{}"}),
-        ):
-            mock_find.return_value.attrs = {"content": "some-token"}
-            result = await test_bot.delete_ad(ad_cfg, published_ads, delete_old_ads_by_title = False)
-
-        assert result is True
-        assert ad_cfg.id is None
-
-    @pytest.mark.asyncio
-    async def test_delete_ad_returns_false_when_no_match(
-        self,
-        test_bot:KleinanzeigenBot,
-        minimal_ad_config:dict[str, Any],
-        caplog:pytest.LogCaptureFixture,
-    ) -> None:
-        """When no published ads match, should return False without opening any pages."""
-        ad_cfg = Ad.model_validate(minimal_ad_config | {"title": "No Match Title", "id": 99999})
-        published_ads = [{"title": "Different Title", "id": 12345}]
-
-        with (
-            caplog.at_level(logging.WARNING, logger = "kleinanzeigen_bot"),
-            patch.object(test_bot, "web_open", new_callable = AsyncMock) as mock_web_open,
-            patch.object(test_bot, "web_find", new_callable = AsyncMock) as mock_web_find,
-            patch.object(test_bot, "web_sleep", new_callable = AsyncMock) as mock_web_sleep,
-            patch.object(test_bot, "web_request", new_callable = AsyncMock) as mock_request,
-        ):
-            result = await test_bot.delete_ad(ad_cfg, published_ads, delete_old_ads_by_title = True)
-
-        assert result is False
-        assert ad_cfg.id == 99999  # Preserved — no deletion attempted
-        mock_web_open.assert_not_called()
-        mock_web_find.assert_not_called()
-        mock_web_sleep.assert_not_called()
-        mock_request.assert_not_called()
-
-    @pytest.mark.asyncio
-    async def test_delete_ad_returns_false_on_404_clears_id(
-        self,
-        test_bot:KleinanzeigenBot,
-        minimal_ad_config:dict[str, Any],
-    ) -> None:
-        """When delete is attempted but server returns 404, should return False but still clear the id."""
-        ad_cfg = Ad.model_validate(minimal_ad_config | {"id": 12345})
-        published_ads:list[dict[str, Any]] = []
-
-        with (
-            patch.object(test_bot, "web_open", new_callable = AsyncMock),
-            patch.object(test_bot, "web_find", new_callable = AsyncMock) as mock_find,
-            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
-            patch.object(test_bot, "web_request", new_callable = AsyncMock, return_value = {"statusCode": 404, "statusMessage": "Not Found", "content": "{}"}),
-        ):
-            mock_find.return_value.attrs = {"content": "some-token"}
-            result = await test_bot.delete_ad(ad_cfg, published_ads, delete_old_ads_by_title = False)
-
-        assert result is False
-        assert ad_cfg.id is None  # Cleared because delete was attempted
-
-    @pytest.mark.asyncio
-    async def test_delete_ad_skips_invalid_published_ad_id(self, test_bot:KleinanzeigenBot, minimal_ad_config:dict[str, Any]) -> None:
-        """When a title-matched published ad has an invalid id (None), it should be skipped."""
-        ad_cfg = Ad.model_validate(minimal_ad_config | {"title": "Test Title", "id": None})
-        published_ads:list[dict[str, Any]] = [
-            {"title": "Test Title", "id": None},  # Invalid — should be skipped
-            {"title": "Test Title", "id": "not-a-number"},  # Invalid — should be skipped
-        ]
-
-        with (
-            patch.object(test_bot, "web_open", new_callable = AsyncMock) as mock_web_open,
-            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
-            patch.object(test_bot, "web_request", new_callable = AsyncMock),
-        ):
-            result = await test_bot.delete_ad(ad_cfg, published_ads, delete_old_ads_by_title = True)
-
-        assert result is False
-        mock_web_open.assert_not_called()  # No valid IDs → no page open
-
-    @pytest.mark.asyncio
-    async def test_delete_ad_with_zero_id(self, test_bot:KleinanzeigenBot, minimal_ad_config:dict[str, Any]) -> None:
-        """When ad_cfg.id is 0 (falsy but valid), should still enter the ID-based deletion path."""
-        ad_cfg = Ad.model_validate(minimal_ad_config | {"id": 0})
-        published_ads:list[dict[str, Any]] = []
-
-        with (
-            patch.object(test_bot, "web_open", new_callable = AsyncMock),
-            patch.object(test_bot, "web_find", new_callable = AsyncMock) as mock_find,
-            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
-            patch.object(test_bot, "web_request", new_callable = AsyncMock,
-                         return_value = {"statusCode": 200, "statusMessage": "OK", "content": "{}"}) as mock_request,
-        ):
-            mock_find.return_value.attrs = {"content": "some-token"}
-            result = await test_bot.delete_ad(ad_cfg, published_ads, delete_old_ads_by_title = False)
-
-        assert result is True
-        assert ad_cfg.id is None
-        mock_request.assert_called_once()
-        assert "ids=0" in mock_request.call_args[1]["url"]
-
-    @pytest.mark.asyncio
-    async def test_delete_ad_multiple_title_matches(self, test_bot:KleinanzeigenBot, minimal_ad_config:dict[str, Any]) -> None:
-        """When multiple published ads match by title with mixed 200/404, should return True."""
-        ad_cfg = Ad.model_validate(minimal_ad_config | {"title": "Test Title", "id": None})
-        published_ads = [
-            {"title": "Test Title", "id": 100},
-            {"title": "Test Title", "id": 200},
-            {"title": "Test Title", "id": 300},
-        ]
-
-        response_sequence = [
-            {"statusCode": 200, "statusMessage": "OK", "content": "{}"},
-            {"statusCode": 404, "statusMessage": "Not Found", "content": "{}"},
-            {"statusCode": 200, "statusMessage": "OK", "content": "{}"},
-        ]
-
-        with (
-            patch.object(test_bot, "web_open", new_callable = AsyncMock),
-            patch.object(test_bot, "web_find", new_callable = AsyncMock) as mock_find,
-            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
-            patch.object(test_bot, "web_request", new_callable = AsyncMock, side_effect = response_sequence) as mock_request,
-        ):
-            mock_find.return_value.attrs = {"content": "some-token"}
-            result = await test_bot.delete_ad(ad_cfg, published_ads, delete_old_ads_by_title = True)
-
-        assert result is True
-        assert ad_cfg.id is None
-        assert mock_request.call_count == 3
-
-    @pytest.mark.asyncio
-    async def test_delete_ad_title_and_id_match_deduplicated(self, test_bot:KleinanzeigenBot, minimal_ad_config:dict[str, Any]) -> None:
-        """When the same ad matches by both title and ID, it should be deleted only once."""
-        ad_cfg = Ad.model_validate(minimal_ad_config | {"title": "Test Title", "id": 100})
-        published_ads = [{"title": "Test Title", "id": 100}]
-        ok_response = {"statusCode": 200, "statusMessage": "OK", "content": "{}"}
-
-        with (
-            patch.object(test_bot, "web_open", new_callable = AsyncMock),
-            patch.object(test_bot, "web_find", new_callable = AsyncMock) as mock_find,
-            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
-            patch.object(test_bot, "web_request", new_callable = AsyncMock, return_value = ok_response) as mock_request,
-        ):
-            mock_find.return_value.attrs = {"content": "some-token"}
-            result = await test_bot.delete_ad(ad_cfg, published_ads, delete_old_ads_by_title = True)
-
-        assert result is True
-        assert ad_cfg.id is None
-        mock_request.assert_called_once()  # Deduplicated — only one request
-
-    @pytest.mark.asyncio
-    async def test_delete_ad_exception_preserves_id(self, test_bot:KleinanzeigenBot, minimal_ad_config:dict[str, Any]) -> None:
-        """When web_request raises an exception mid-loop, ad_cfg.id should be preserved and web_sleep not called."""
-        ad_cfg = Ad.model_validate(minimal_ad_config | {"id": 12345})
-        published_ads:list[dict[str, Any]] = []
-
-        with (
-            patch.object(test_bot, "web_open", new_callable = AsyncMock),
-            patch.object(test_bot, "web_find", new_callable = AsyncMock) as mock_find,
-            patch.object(test_bot, "web_sleep", new_callable = AsyncMock) as mock_web_sleep,
-            patch.object(test_bot, "web_request", new_callable = AsyncMock, side_effect = TimeoutError("request timed out")),
-        ):
-            mock_find.return_value.attrs = {"content": "some-token"}
-            with pytest.raises(TimeoutError, match = "request timed out"):
-                await test_bot.delete_ad(ad_cfg, published_ads, delete_old_ads_by_title = False)
-
-        assert ad_cfg.id == 12345  # Preserved — exception prevented clearing
-        mock_web_sleep.assert_not_called()
 
 
 class TestKleinanzeigenBotShippingOptions:
@@ -2738,7 +2076,7 @@ class TestKleinanzeigenBotShippingOptions:
             patch("kleinanzeigen_bot.Path", PureWindowsPath),
             patch("kleinanzeigen_bot.price_reduction.apply_auto_price_reduction", side_effect = mock_apply_auto_price_reduction),
             patch.object(test_bot, "web_open", new_callable = AsyncMock),
-            patch.object(test_bot, "delete_ad", new_callable = AsyncMock),
+            patch("kleinanzeigen_bot.delete_flow.delete_ad", new_callable = AsyncMock),
         ):
             # Call publish_ad and expect sentinel exception
             try:
@@ -4993,59 +4331,3 @@ class TestTrackingFallback:
             result = await test_bot._try_recover_ad_id_from_redirect()
 
         assert result == 11223344
-
-
-class TestDeleteAdsAfterDeletePolicy:
-    """Tests for delete_ads orchestration with after_delete policy integration."""
-
-    @staticmethod
-    def _make_ad(minimal_ad_config:dict[str, Any], tmp_path:Path) -> tuple[str, Ad, dict[str, Any]]:
-        ad_cfg = Ad.model_validate(minimal_ad_config | {
-            "id": 12345, "active": True,
-            "created_on": "2024-06-01T12:00:00", "updated_on": "2024-06-10T08:30:00",
-            "content_hash": "abc123", "repost_count": 3, "price_reduction_count": 1,
-        })
-        return str(tmp_path / "ad.yaml"), ad_cfg, ad_cfg.model_dump()
-
-    @pytest.mark.asyncio
-    async def test_cleanup_on_404_detection(
-        self, test_bot:KleinanzeigenBot, minimal_ad_config:dict[str, Any], tmp_path:Path,
-    ) -> None:
-        """Cleanup runs when delete_ad returns False but cleared the id (404 path)."""
-        test_bot.config.deleting.after_delete = "RESET"
-        ad_file, ad_cfg, ad_cfg_orig = self._make_ad(minimal_ad_config, tmp_path)
-
-        async def fake_delete(ad:Ad, _published:list[dict[str, Any]], **__:Any) -> bool:
-            ad.id = None  # Phase B ran and cleared the id
-            return False  # but all responses were 404
-
-        with (
-            patch.object(test_bot, "_fetch_published_ads", new_callable = AsyncMock, return_value = []),
-            patch.object(test_bot, "delete_ad", new_callable = AsyncMock, side_effect = fake_delete),
-            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
-            patch("kleinanzeigen_bot.utils.dicts.save_dict") as mock_save,
-        ):
-            await test_bot.delete_ads([(ad_file, ad_cfg, ad_cfg_orig)])
-
-        assert ad_cfg.repost_count == 0
-        assert "id" not in ad_cfg_orig
-        mock_save.assert_called_once()
-
-    @pytest.mark.asyncio
-    async def test_no_cleanup_when_delete_not_attempted(
-        self, test_bot:KleinanzeigenBot, minimal_ad_config:dict[str, Any], tmp_path:Path,
-    ) -> None:
-        """No cleanup when delete_ad returns False with id preserved (no match)."""
-        test_bot.config.deleting.after_delete = "RESET"
-        ad_file, ad_cfg, ad_cfg_orig = self._make_ad(minimal_ad_config, tmp_path)
-
-        with (
-            patch.object(test_bot, "_fetch_published_ads", new_callable = AsyncMock, return_value = []),
-            patch.object(test_bot, "delete_ad", new_callable = AsyncMock, return_value = False),
-            patch.object(test_bot, "web_sleep", new_callable = AsyncMock),
-            patch("kleinanzeigen_bot.utils.dicts.save_dict") as mock_save,
-        ):
-            await test_bot.delete_ads([(ad_file, ad_cfg, ad_cfg_orig)])
-
-        mock_save.assert_not_called()
-        assert ad_cfg.id == 12345
